LabaRide/advancedb_project/backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
mysql-connector-python==8.0.26
bcrypt==3.2.0
python-dotenv==0.19.0
python-jose==3.3.0